fix: siren size from 16 bits to 32

+++ src/api/auth.rs
@@ -54,7 +54,7 @@ pub struct RegisterRequest {
     pub name: String,
     pub password: String,
     pub role: Role,
-    pub siren: Option<i16>,
+    pub siren: Option<i32>,
 }
 
 #[post("/register")]
